feat(bookstore): create GET endpoint with path parameter to get specific book based on title
With the path parameter we can get a specific book based on the title. The path parameter is defined within curly braces.

With FastAPI we can write directly this parameter as the function argument and add type hint. Because FastAPI works with pydantic behind scenes, if we passed a parameter that is not a string we will get an error.