feat(bookstore): create POST endpoint to create a new book.
The POST operation is intended for those operations where we want to create new data, in this case, a new book. By default the new book is set to be the Body, which is the data the client passes within the call to the endpoint.

By adding type hint we avoid the user to introduce something that is not a dictionary of strings, but right now we can't do anything if the client passes a dictionary with different keys and values