feat(bookstore): create PUT endpoint to update book data
For PUT operations we can pass a body as we've done in the POST endpoint. This body will be the updated values of the book we want to update based on the title of the book.

As we mentioned with the POST endpoint. The type hint prevent the client to pass something that is not a dictionary, but we don't have a way of preventing the client to use different keys