feat(assignment): create GET endpoint to filter books by author using path or query parameters.
In this case with query parameters we need to be aware of the order of the endpoints. We should always put the smaller and static endpoints first to avoid calling the incorrect endpoint.

In this assignment, if the /books/by_author/ endpoint were at the end of the file, when the endpoint is called it would be redirected to the /books/{author}/ endpoint because it has the same structure.