feat(advanced-bookstore): create basic GET endpoint to return all the books
Now our books variable is a list of Book objects instead of a list of dictionaries. Until we see how pydantic and its models work, we won't be adding a return type hint (if we write list[Book] pydantic raises an error).

One advantage of using FastAPI is that it convert automatically the the response to a json format, so when we return books, the client will received a list of json, where each json has the id, title, author etc.