refactor(advanced-bookstore): add pydantic model to POST argument
by adding the type hint to the book argument we allow FastAPI and pydantic to validate automatically the request based on the BookRequest class we've defined previously.

To create a new Book instance from the BookRequest we need to use the model_dump method, this will convert the fields to a dictionary so we can create correctly the Book object.