refactor(advanced-bookstore): add return type hint to GET endpoints to validate outputs
In cases where we are returning some data that is not exactly what the pydantic model declares we can use the response_model parameter in the decorator. This argument will allow us to still generate the documentation, filter the fields from the Book that are not in the BookResponse and validate its values.

We could add the return type hint as well, writing list[BookResponse] instead of Any, but then the editor will throw a warning because the return type hint and the returned object are not the same.