refactor(advanced-bookstore): add status codes to endpoints when the request is successfully managed.
We can set the status code that will be send to the client when the operation is successful using the status_code argument inside the endpoint decorator.

Normally, for GET methods we return a 200 to say everything is okay, for POST methods we return a 201 to inform the resource was created successfully and 204 for PUT operations when the data is updated correctly. Careful, if we defined a 204 status code we can't return anything.