refactor(todos-router): user must be authenticated to be able to create a task
When the user calls the endpoint, it will execute the get_authenticated_user. At the same time, this function depends on the login endpoint, so if the user is not authenticated it wont let them to create a task.

In the swagger UI it will appear a lock on the right side of the endpoint title, this means that the endpoint needs authentication. We can authenticate by clicking on that lock and filling the fields. In this case what we would be calling is the login endpoint