refactor(model): remove index from all columns of the tables
Setting all columns to be indexed is not a good practice as it will consume storage in disk. By default columns marked as unique and primary keys will be automatically indexed to be able to know its uniqueness quickly. Apart from these, we would want to set a column to have an index only if it's common to do search operation on that field